requirements.txt
google-api-python-client>=2.90.0
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.8.0
google-cloud-aiplatform>=1.30.0
google-api-core>=2.28.1
requests>=2.31.0
python-dotenv>=1.0.0
